requirements.txt
torch==1.8.1
torchvision==0.9.1
pytorch-lightning==1.2.10
ray==1.3.0
pillow
matplotlib
pandas
simpleCRF
albumentations
scikit-image
tabulate==0.8.9
